requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-cors-headers==4.2.0
python-decouple==3.8
psycopg2-binary==2.9.7
djangorestframework-simplejwt==5.2.2